Feature 2025 06 26

+++ EducacaoXpert/src/EducacaoXpert.Core/DomainObjects/DTO/CursoDto.cs
@@ -4,6 +4,6 @@ public class CursoDto
 {
     public Guid Id { get; set; }
     public string Nome { get; set; }
-    public decimal Preco { get; set; }
+    public int Preco { get; set; }
     public List<AulaDto> Aulas { get; set; }
 }
